removing confusing html character codes
This is motivated by the fact that &#8209; is actually a different symbol from - (the minus sign on my kbd). This causes problems for simple find (Ctrl-f) in chrome, as searching for --match isnt the same as searching for ‑‑match! This could additionally cause problems when examples are coppied and pasted from the html page, i.e. "ls ‑l" gives "ls: cannot access ‑l: No such file or directory", hahahaha. Since I was getting rid of 98% the codes anyway, I went ahead and removed the other 2% for consistency as well. I imagine they just got used by a copy paste from a text editor on a mac or so...